fix(invoicing): batch invoice 500'd, and the coverage ledger behind it was empty (#1352)
Heritage Condominium's "Consolidate 4" returned "Failed to generate invoice".
Two separate defects from #1349, both in the new billing-coverage engine.

1. Drizzle does not bind a JS array through ANY().
   `= ANY(${ids}::int[])` spreads the elements into separate placeholders:
   one id emits ANY(($2)::int[]) -> malformed array literal: "9099";
   two emit ANY(($2, $3)::int[]) -> cannot cast type record to integer[].
   The pattern throws at every length. setInvoiceCoverage hit the
   one-element form for each per-visit invoice, and the exception surfaced
   as the toast. Replaced all 11 sites in invoice-billing.ts with a
   parameterized IN list (one placeholder per value, values coerced to
   numbers). Same trap already documented at users.ts:189 and sms.ts:379.

2. The invoice_job_links backfill has never once succeeded.
   A bare NULL in its SELECT list types as text, so Postgres rejected the
   insert into a numeric column. The per-statement catch logged it
   "non-fatal" and boot continued, so the table sat empty company-wide
   (0 rows, 1150 expected) while the uninvoiced-jobs queue, the Generate
   Invoice button and the completion engine had already switched to
   reading it as THE dedup predicate. Every already-invoiced visit read as
   unbilled. Cast to NULL::numeric.

Defect 2 is why defect 1 could not be fixed on its own: with an empty
ledger, a working button would have billed Heritage's Aug 3 visit a second
time -- invoice 7240 already covers it, sent, $108.25. That is the KMA
double-bill the #1349 design exists to prevent.

Verified against production, every write rolled back:
  - old form reproduces the exact logged error for job 9099
  - all 11 fixed statements execute at 1 and 2 ids
  - backfill inserts 1150 links, promotes 985 live, 0 visits with >1 live
  - job 9099 resolves to invoice 7240; Heritage's uninvoiced count goes
    4 -> 3, dropping only the already-billed visit

+++ artifacts/api-server/src/phes-data-migration.ts
@@ -2117,9 +2117,19 @@ async function runBookingSchemaGuard(): Promise<void> {
     // unique index can never be violated mid-backfill.
     // The `~ '^[0-9]+$'` guard before ::int is load-bearing — a bare cast over
     // mixed jsonb 500s the whole statement (the #1146 comm-log regression).
+    // The `NULL::numeric` cast is load-bearing. A bare NULL in a SELECT list is
+    // type `unknown`, which Postgres resolves to `text` here and then refuses:
+    // "column amount is of type numeric but expression is of type text". The
+    // per-statement catch below logged that as non-fatal and carried on, so this
+    // backfill threw on EVERY boot since it shipped and invoice_job_links stayed
+    // empty company-wide — while the uninvoiced-jobs queue, the Generate Invoice
+    // button and the completion engine had already switched to reading it as
+    // THE dedup predicate. An empty ledger means every already-invoiced visit
+    // reads as unbilled, which is the KMA double-bill re-armed. Verified on
+    // production 2026-08-03: 0 rows present, 1150 expected.
     { label: "backfill invoice_job_links", stmt: `
       INSERT INTO invoice_job_links (company_id, invoice_id, job_id, amount, is_live)
-      SELECT DISTINCT i.company_id, i.id, cov.job_id, NULL, FALSE
+      SELECT DISTINCT i.company_id, i.id, cov.job_id, NULL::numeric, FALSE
         FROM invoices i
         CROSS JOIN LATERAL (
           SELECT i.job_id AS job_id WHERE i.job_id IS NOT NULL
